+++ cli.py
@@ -43,9 +43,10 @@ def cli() -> None:
 @cli.command()
 @click.option('-p', '--port', default=Config.DEFAULT_PORT, help='The port to bind to.')
 @click.option('-x', '--prefix', default=Config.DEFAULT_INTERNAL_PREFIX, help='Url prefix of internal endpoints.')
-def run(port: int, prefix: str) -> None:
+@click.option('-r', '--routes', type=click.Path(exists=True), help='Path to json file with default routes.')
+def run(port: int, prefix: str, default_routes: str) -> None:
     """Start local Trickster app."""
-    config = Config(internal_prefix=prefix, port=port)
+    config = Config(internal_prefix=prefix, port=port, routes_path=default_routes)
     app = ApiApp(config)
     app.run()